+++ config/target/prod.exs
@@ -35,8 +35,14 @@ config :farmbot, :init, [
   # Start up Network
   Farmbot.Target.Network,
 
+  # SSH Console.
+  Farmbot.Target.SSHConsole,
+
   # Wait for time time come up.
   Farmbot.Target.Network.WaitForTime,
+
+  # Wait for DNS resolution
+  Farmbot.Target.Network.DnsTask,
 
   # Stops the disk from getting full.
   Farmbot.Target.Network.TzdataTask,
@@ -65,7 +71,8 @@ config :farmbot, :behaviour,
   pin_binding_handler: Farmbot.Target.PinBinding.AleHandler,
   leds_handler: Farmbot.Target.Leds.AleHandler
 
+config :nerves_network, regulatory_domain: "US"
 config :shoehorn,
-  init: [:nerves_runtime],
+  init: [:nerves_runtime, :nerves_firmware_ssh],
   handler: Farmbot.ShoehornHandler,
   app: :farmbot