+++ lib/farmbot/bootstrap/settings_sync.ex
@@ -143,7 +143,12 @@ defmodule Farmbot.Bootstrap.SettingsSync do
   end
 
   @doc false
-  def run(tries_left \\ 6) do
+  def run(tries_left \\ 6)
+  def run(0) do
+    Logger.error 1, "failed to sync settings"
+  end
+
+  def run(tries_left) do
     # HACK: Connor - Don't even look at this.
     if is_nil(Process.whereis(Farmbot.System.ConfigStorage)) do
       Logger.error 1, "ConfigStorage not running? trying #{tries_left - 1} more times"
@@ -151,8 +156,15 @@ defmodule Farmbot.Bootstrap.SettingsSync do
       run(tries_left - 1)
     else
       try do
-        do_sync_fw_configs()
-        do_sync_fbos_configs()
+        update_config_value(:bool, "settings", "ignore_fbos_config", true)
+        update_config_value(:bool, "settings", "ignore_fw_config", true)
+
+          do_sync_fw_configs()
+          do_sync_fbos_configs()
+
+        update_config_value(:bool, "settings", "ignore_fbos_config", false)
+        update_config_value(:bool, "settings", "ignore_fw_config", false)
+
         Logger.debug 1, "Synced Farmbot OS and Firmware settings with API"
         :ok
       rescue
@@ -161,6 +173,7 @@ defmodule Farmbot.Bootstrap.SettingsSync do
           err_msg = "#{message} #{inspect System.stacktrace()}"
           Logger.error 1, "Error syncing settings: #{err_msg}"
           update_config_value(:bool, "settings", "ignore_fbos_config", false)
+          update_config_value(:bool, "settings", "ignore_fw_config", false)
           {:error, message}
       end
     end
@@ -196,6 +209,7 @@ defmodule Farmbot.Bootstrap.SettingsSync do
         if old_map[param] != new_value do
           IO.puts "2 #{param} #{old_map[param]} != #{new_value}"
           Farmbot.Firmware.update_param(String.to_atom(param), new_value)
+          Farmbot.Firmware.read_param(String.to_atom(param))
         end
         {param, get_config_value(:float, "hardware_params", param)}
       end)
@@ -257,32 +271,12 @@ defmodule Farmbot.Bootstrap.SettingsSync do
     end
   end
 
-  def do_sync_fw_configs(%{"api_migrated" => true} = api_data) do
-    Logger.info 3, "API is the source of truth for Firmware configs. Downloading data."
-    old_config = get_config_as_map()["hardware_params"]
-    apply_fw_map(old_config, api_data)
-    :ok
-  end
-
-  def do_sync_fw_configs(%{"api_migrated" => false} = api_data) do
-    current = get_config_as_map()["hardware_params"]
-    if current["param_version"] do
-      Logger.info 3, "Farmbot is the source of truth for Firmware configs. Uploading data."
-      payload = Map.put(current, "api_migrated", true) |> Poison.encode!()
-      Farmbot.HTTP.delete!("/api/firmware_config")
-      Farmbot.HTTP.put!("/api/firmware_config", payload)
-    else
-      Logger.info 3, "API has never migrated Firmware configs " <>
-      "But Farmbot has no valid paramaters. Using defaults."
-      payload = Map.put(api_data, "api_migrated", true) |> Poison.encode!()
-      Farmbot.HTTP.put!("/api/firmware_config", payload)
+  def do_sync_fw_configs(api_data) do
+    update_config_value(:bool, "settings", "ignore_fw_config", true)
+      Logger.info 3, "API is source of truth for fw configs."
+      current = get_config_as_map()["hardware_params"]
       apply_fw_map(current, api_data)
-    end
-    :ok
-  end
-
-  def do_sync_fw_configs(_) do
-    Logger.error 1, "API is out of date. Not syncing firmware settings!"
+    update_config_value(:bool, "settings", "ignore_fw_config", false)
     :ok
   end
 
@@ -300,7 +294,8 @@ defmodule Farmbot.Bootstrap.SettingsSync do
     with {:ok, %{body: body, status_code: 200}} <- Farmbot.HTTP.get("/api/fbos_config"),
     {:ok, data} <- Poison.decode(body)
     do
-      do_sync_fbos_configs(data)
+      hacked_data = %{data | "firmware_hardware" => get_config_value(:string, "settings", "firmware_hardware")}
+      do_sync_fbos_configs(hacked_data)
     else
       {:ok, status_code: code} ->
         Logger.error 1, "HTTP error syncing settings: #{code}"
@@ -311,46 +306,13 @@ defmodule Farmbot.Bootstrap.SettingsSync do
     end
   end
 
-  def do_sync_fbos_configs(%{"api_migrated" => true} = api_data) do
-    Logger.info 3, "API is the source of truth for Farmbot OS configs. Downloading data."
-    old_config = get_config_as_map()["settings"]
-    apply_fbos_map(old_config, api_data)
-    :ok
-  end
-
-  def do_sync_fbos_configs(_unimportant_data) do
-    Logger.info 3, "Farmbot is the source of truth for Farmbot OS and Firmware configs. Uploading data."
+  def do_sync_fbos_configs(api_data) do
     update_config_value(:bool, "settings", "ignore_fbos_config", true)
-    auto_sync = get_config_value(:bool, "settings", "auto_sync")
-    beta_opt_in = get_config_value(:bool, "settings", "beta_opt_in")
-    disable_factory_reset = get_config_value(:bool, "settings", "disable_factory_reset")
-    firmware_output_log = get_config_value(:bool, "settings", "firmware_output_log")
-    firmware_input_log = get_config_value(:bool, "settings", "firmware_input_log")
-    sequence_body_log = get_config_value(:bool, "settings", "sequence_body_log")
-    sequence_complete_log = get_config_value(:bool, "settings", "sequence_complete_log")
-    sequence_init_log = get_config_value(:bool, "settings", "sequence_init_log")
-    arduino_debug_messages = get_config_value(:bool, "settings", "arduino_debug_messages")
-    os_auto_update = get_config_value(:bool, "settings", "os_auto_update")
-    firmware_hardware = get_config_value(:string, "settings", "firmware_hardware")
-    network_not_found_timer = get_config_value(:float, "settings", "network_not_found_timer")
-    payload = %{
-      api_migrated: true,
-      auto_sync: auto_sync,
-      beta_opt_in: beta_opt_in,
-      disable_factory_reset: disable_factory_reset,
-      firmware_output_log: firmware_output_log,
-      firmware_input_log: firmware_input_log,
-      sequence_body_log: sequence_body_log,
-      sequence_complete_log: sequence_complete_log,
-      sequence_init_log: sequence_init_log,
-      arduino_debug_messages: arduino_debug_messages,
-      os_auto_update: os_auto_update,
-      firmware_hardware: firmware_hardware,
-      network_not_found_timer: network_not_found_timer,
-    } |> Poison.encode!()
-    Farmbot.HTTP.delete!("/api/fbos_config")
-    Farmbot.HTTP.put!("/api/fbos_config", payload)
+      Logger.info 3, "API is the source of truth for Farmbot OS configs. Downloading data."
+      old_config = get_config_as_map()["settings"]
+      apply_fbos_map(old_config, api_data)
     update_config_value(:bool, "settings", "ignore_fbos_config", false)
+
     :ok
   end