+++ priv/config_storage/migrations/20180814202716_add_ignore_fw_config.exs
@@ -1,0 +1,8 @@
+defmodule Farmbot.System.ConfigStorage.Migrations.AddIgnoreFwConfig do
+  use Ecto.Migration
+  import Farmbot.System.ConfigStorage.MigrationHelpers
+
+  def change do
+    create_settings_config("ignore_fw_config", :bool, false)
+  end
+end